[refactor] 候補の選択をリファクタリング
ArtifactSpoilerとMonsterSpoilerの両方で行っている
候補を表示してリアクションで選択する機能を、
CandidatesSelectorモジュールに共通化した。
